requirements.txt
python-dotenv~=1.0.0
fastapi~=0.108.0
aiofiles~=23.2.1
starlette~=0.32.0.post1
pydantic~=2.5.3
passlib~=1.7.4
tortoise-orm
tortoise-orm[asyncpg]
aerich
avatar-generator
python-jose[cryptography]
python-multipart
uvicorn